Combine auth and get modules and add Auth component
